# .docs/openapi.yaml
openapi: 3.0.3
info:
  title: Irori Chat API
  description: |
    招待制の囲炉裏型チャットアプリ「Irori」のバックエンド REST API および WebSocket 仕様書。
    
    ### 認証方式
    初期開発・ローカル検証向けに簡易認証を適用しています。
    リクエストヘッダーに `Authorization: Bearer <userId>` を付与するか、Cookie に `userId=<userId>` を設定してください。
    指定された `userId` がデータベースに存在しない場合、または指定がない場合、ゲストユーザー（`Guest_xxxx`）が自動生成されます。
  version: 0.1.0
servers:
  - url: http://localhost:8787
    description: ローカル開発環境 (wrangler dev)
paths:
...
  /api/v1/friends:
    get:
      summary: つながっているフレンド一覧取得
      description: ログインユーザーのフレンド一覧を取得します。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  friends:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /api/v1/friends/invite-code:
    post:
      summary: フレンド招待コードの生成
      description: 新たに他のユーザーをフレンド追加するためのランダムな招待コードを生成します。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  inviteCode:
                    type: string
                    format: uuid
                    example: "c3b9a71a-289c-482d-a2f1-081e64906f3a"
  /api/v1/friends/connect:
    post:
      summary: 招待コードを用いたフレンド登録
      description: 指定された招待コードを照合し、招待者と自身をフレンドとして登録（相互接続）します。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inviteCode
              properties:
                inviteCode:
                  type: string
                  format: uuid
                  example: "c3b9a71a-289c-482d-a2f1-081e64906f3a"
      responses:
        '200':
          description: 接続成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: 自分自身への接続などのエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 招待コードが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/rooms/{roomId}/messages:
    get:
      summary: ルームの過去メッセージ一覧（タイムライン）取得
      description: |
        ルーム所属メンバーのみ取得可能です。
        一時チャットルーム（`isEphemeral: true`）の場合は、自動的に過去24時間以内のメッセージのみにフィルタリングされます。
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          description: ページネーション用のカーソル（最後に取得したメッセージの createdAt 文字列）。
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: 取得件数 (最大100)
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  nextCursor:
                    type: string
                    nullable: true
                    example: "2026-07-17T12:00:00.000Z"
        '403':
          description: メンバーではないためアクセス不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: 新規メッセージの投稿
      description: 投稿成功時、該当ルームの Durable Object へ通信が走り、リアルタイムに接続中のクライアントへ `NEW_MESSAGE` イベントがブロードキャストされます。
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "今夜、囲炉裏で話しましょう"
                mediaUrl:
                  type: string
                  nullable: true
                  example: "https://drive.google.com/file/d/.../view"
                replyToId:
                  type: string
                  nullable: true
                  example: "m_12345"
      responses:
        '201':
          description: 投稿成功・保存完了
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: "5a2f6d1c-b84a-4a7b-a2f1-081e64908d1c"
                  status:
                    type: string
                    example: "saved"
        '403':
          description: メンバーではないため投稿不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/messages/{messageId}:
    put:
      summary: メッセージの編集
      description: 投稿者本人のみが編集可能です。更新成功時、Durable Object 経由で接続クライアントへ `MESSAGE_EDIT` がブロードキャストされます。
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "編集後のメッセージ"
      responses:
        '200':
          description: 編集成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '403':
          description: 投稿者本人ではないため編集不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: メッセージが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: メッセージの削除（送信取り消し）
      description: 投稿者本人、もしくはそのルームの管理者 (role=admin) のみが削除可能です。削除成功時、Durable Object 経由で接続クライアントへ `MESSAGE_DELETE` がブロードキャストされます。
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '403':
          description: 削除権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: メッセージが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 認証用の `userId` 文字列を直接設定します。
  schemas:
    StatusLamp:
      type: string
      enum:
        - free
        - busy
        - away
      example: free
    User:
      type: object
      properties:
        id:
          type: string
          example: "u123"
        name:
          type: string
          example: "Guest_e123"
        statusLamp:
          $ref: '#/components/schemas/StatusLamp'
...
    Message:
      type: object
      properties:
        id:
          type: string
          example: "m_12345"
        roomId:
          type: string
          format: uuid
          example: "8f7e6d1c-b84a-4a7b-a2f1-081e64908d1c"
        userId:
          type: string
          nullable: true
          description: 投稿したユーザーのID。Webhook等のシステム投稿の場合は null になります。
          example: "u123"
        content:
          type: string
          example: "こんにちは"
        mediaUrl:
          type: string
          nullable: true
          example: "https://drive.google.com/file/d/.../view"
        replyToId:
          type: string
          nullable: true
          example: "m_10000"
        createdAt:
          type: string
          format: date-time
          example: "2026-07-17T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-17T12:00:00.000Z"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Forbidden"